Add files via upload
I kept getting an error in the terminal window when I first used the  " git add -A" and  " git commit -m 'created game logic' " and " git push origin master " command. so I wasn't sure if my work was saved in the wdi-fundamentals-memorygame repository. So I added my files and draged them here, and I also attached a screenshot of my terminal window to show you the error I was getting.